requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework
psycopg2-binary
pandas
pytz